+++ packages/go_router/test/builder_test.dart
@@ -7,7 +7,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:go_router/src/builder.dart';
 import 'package:go_router/src/configuration.dart';
 import 'package:go_router/src/match.dart';
-import 'package:go_router/src/matching.dart';
 import 'package:go_router/src/router.dart';
 
 import 'test_helpers.dart';
@@ -36,8 +35,6 @@ void main() {
             RouteMatch(
               route: config.routes.first as GoRoute,
               matchedLocation: '/',
-              extra: null,
-              error: null,
               pageKey: const ValueKey<String>('/'),
             ),
           ],
@@ -121,8 +118,6 @@ void main() {
             RouteMatch(
               route: config.routes.first as GoRoute,
               matchedLocation: '/',
-              extra: null,
-              error: null,
               pageKey: const ValueKey<String>('/'),
             ),
           ],
@@ -176,15 +171,11 @@ void main() {
             RouteMatch(
               route: config.routes.first,
               matchedLocation: '',
-              extra: null,
-              error: null,
               pageKey: const ValueKey<String>(''),
             ),
             RouteMatch(
               route: config.routes.first.routes.first,
               matchedLocation: '/details',
-              extra: null,
-              error: null,
               pageKey: const ValueKey<String>('/details'),
             ),
           ],
@@ -251,8 +242,6 @@ void main() {
             RouteMatch(
               route: config.routes.first.routes.first as GoRoute,
               matchedLocation: '/a/details',
-              extra: null,
-              error: null,
               pageKey: const ValueKey<String>('/a/details'),
             ),
           ],